refactor(dto): 금액을 int 가 아닌 BigDecimal 로 받도록 변경

+++ src/main/java/roomescape/payment/dto/PaymentCreateRequest.java
@@ -1,8 +1,9 @@
 package roomescape.payment.dto;
 
+import java.math.BigDecimal;
 import roomescape.reservation.domain.Reservation;
 
-public record PaymentCreateRequest(String paymentKey, String orderId, int amount, Reservation Reservation) {
+public record PaymentCreateRequest(String paymentKey, String orderId, BigDecimal amount, Reservation Reservation) {
     public RestClientPaymentApproveRequest createRestClientPaymentApproveRequest() {
         return new RestClientPaymentApproveRequest(paymentKey, orderId, amount);
     }
