+++ src/main/java/roomescape/payment/dto/RestClientPaymentApproveRequest.java
@@ -1,4 +1,6 @@
 package roomescape.payment.dto;
 
-public record RestClientPaymentApproveRequest(String paymentKey, String orderId, int amount) {
+import java.math.BigDecimal;
+
+public record RestClientPaymentApproveRequest(String paymentKey, String orderId, BigDecimal amount) {
 }
